% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMut.R
\name{countMut}
\alias{countMut}
\title{Returns gene name and frequency of each gene types}
\usage{
countMut(mutFile, hgncFile)
}
\arguments{
\item{mutFile}{A .tsv file containing COSMIC point mutations}

\item{hgncFile}{A .tsv file with complete approved human gene nomenclature IDs}
}
\value{
Returns dataframe with gene names and frequency.
}
\description{
counMut function uses cosImport function in cosImport.R file to read from
cosmic file and uses that dataframe to extract IDsymbol column and return the
occurrence frequency of each gene types.
}
\examples{
# Using data files available in the package: "MutantData.tsv" and "HGNC.tsv"

# countMut("Data/MutantData.tsv", "Data/HGNC.tsv")

}
