% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cds.R
\name{plot_cds}
\alias{plot_cds}
\title{Returns barplot of mutation (cds) occurences}
\usage{
plot_cds(mutFile, hgncFile)
}
\arguments{
\item{mutFile}{A .tsv file containing COSMIC point mutations}

\item{hgncFile}{A .tsv file with complete approved human gene nomenclature IDs}
}
\value{
Returns barplot (descending order) using mutation frequency data
obtained from mut_cds function
}
\description{
plot_cds function uses mut_cds function in mut_cds.R file to get frequency data
of each mutation and return barplot using the data extracted
}
\examples{
# Using data files available in the package: "MutantData.tsv" and "HGNC.tsv"

# plot_cds("Data/MutantData.tsv", "Data/HGNC.tsv")

}
\references{
Kabacoff , R. (n.d.). Bar plots. Quick-R: Bar Plots. Retrieved November 21,
2021, from https://www.statmethods.net/graphs/bar.html.

Abedin, J., &amp; Mittal, H. V. (n.d.). R Graphs. R Graphs Cookbook (Second
Edition) - Displaying values on top of or next to the bars. Retrieved November
 21, 2021, from https://subscription.packtpub.com/book/big_data_and_business_
 intelligence/9781783988785/6/ch06lvl1sec69/displaying-values-on-top-of-or-
 next-to-the-bars.

  @export
}
