% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCANMutVisual.R
\name{runCANMutVisual}
\alias{runCANMutVisual}
\title{Launch Shiny App for package CANMutVisual}
\usage{
runCANMutVisual()
}
\value{
The function will launch the shiny app. No return value
}
\description{
The shiny app allow user to choose two input files to process
the functions provided in the package. The visualization using Shiny app
help users to utilize this package with no prior knowledge on R.
}
\examples{
 \dontrun{
 CanMutVisual::runCanMutVisual()
 }

}
\references{
https://shiny.rstudio.com/tutorial/
}
